+++ docs/changes.rst
@@ -1,12 +1,21 @@
 Changelog
 =========
+
+3.0.0 (2020-11-)
+----------------
+* Rework of executors
+    * `explore` is now a method on the executor object
+    * Add a `man`-page executor
+    * `aclimatise.explore` is now deprecated
+
+
 2.2.0 (2020-09-26)
 ------------------
 * Add ``command.reanalyse()``, which re-analyses the help text stored in the command, using the current parser
 * Add ``command['sort']`` which lets you access subcommands by name
 
 2.1.0 (2020-09-22)
-----------------
+------------------
 * Add `bedtools random` as a test case
 * Use a simple metric to exclude flags unless they have a somewhat valid description text, this should result in less false-positive flags being detected
 * Fixed the `id` field in the CWL tool definitions including some path elements (like `..`)
@@ -19,9 +28,9 @@ Changelog
 * Rename the CLI utility from ``acclimatise`` to ``aclimatise``.
 
 1.2.0 (2020-09-13)
------------------
+------------------
 User-Facing Changes
-******************
+*******************
 * Add output generation for CWL and WDL
 
 Fixes